+++ packages/volto-light-theme/src/config/blocks.tsx
@@ -39,6 +39,7 @@ import { tocBlockSchemaEnhancer } from '../components/Blocks/Toc/schema';
 import { mapsBlockSchemaEnhancer } from '../components/Blocks/Maps/schema';
 import { sliderBlockSchemaEnhancer } from '../components/Blocks/Slider/schema';
 import EventMetadataView from '../components/Blocks/EventMetadata/View';
+import isEmpty from 'lodash/isEmpty';
 
 declare module '@plone/types' {
   export interface BlocksConfigData {
@@ -131,15 +132,23 @@ export default function install(config: ConfigType) {
     },
   ];
 
-  function blockThemesEnhancer(data) {
+  function blockThemesEnhancer({ data, container }) {
     if (!data['@type']) return {};
+    const blockConfig = config.blocks.blocksConfig[data['@type']];
+    const blockStyleDefinitions =
+      // We look up for the blockThemes in the block's config, then in the global config
+      // We keep `colors` for BBB, but `themes` should be used
+      blockConfig.themes || blockConfig.colors || config.blocks.themes || [];
+    if (!blockConfig) return {};
+
+    if (
+      !isEmpty(container) &&
+      container.theme &&
+      (!data.theme || data.theme === 'default')
+    ) {
+      return findStyleByName(blockStyleDefinitions, container.theme);
+    }
     if (data.theme) {
-      const blockConfig = config.blocks.blocksConfig[data['@type']];
-      if (!blockConfig) return {};
-      const blockStyleDefinitions =
-        // We look up for the blockThemes in the block's config, then in the global config
-        // We keep `colors` for BBB, but `themes` should be used
-        blockConfig.themes || blockConfig.colors || config.blocks.themes || [];
       return data.theme
         ? findStyleByName(blockStyleDefinitions, data.theme)
         : {};
@@ -155,11 +164,11 @@ export default function install(config: ConfigType) {
     method: blockThemesEnhancer,
   });
 
-  // No required blocks (eg. Title)
+  // No required blocks except eventMetadata
   config.blocks.requiredBlocks = [
     ...config.blocks.requiredBlocks,
     'eventMetadata',
-  ];
+  ].filter((block) => block !== 'title');
 
   // Remove Hero Block (No longer in Volto 18)
   // TODO: Remove already?