+++ packages/volto-light-theme/src/customizations/volto/components/theme/View/RenderBlocks.jsx
@@ -88,7 +88,7 @@ const RenderBlocks = (props) => {
         const themes =
           config.blocks.blocksConfig[
             content[blocksFieldname][group[0]]['@type']
-          ].themes ?? config.blocks.themes;
+          ]?.themes ?? config.blocks.themes;
 
         return (
           <MaybeWrap
@@ -145,6 +145,7 @@ const RenderBlocks = (props) => {
                       id={block}
                       block={block}
                       data={blockData}
+                      isContainer={isContainer}
                     >
                       <Block
                         id={block}